v2.1 1.后续改成uuid作为token，token作为key存入redis，redis存入具体的用户信息。后续获取用户信息从redis中获取。

+++ user/src/main/java/com/pxy/user/mapper/MenuMapper.java
@@ -10,5 +10,5 @@ public interface MenuMapper {
      * @param userId
      * @return
      */
-    public List<Menu> getMenus(Integer userId);
+    public List<Menu> getMenus(Long userId);
 }
